fix(migration): make misc migrations idempotent

+++ lib/Migration/Version4100Date20241021091352.php
@@ -22,8 +22,13 @@ class Version4100Date20241021091352 extends SimpleMigrationStep {
 	 * @return null|ISchemaWrapper
 	 */
 	public function changeSchema(IOutput $output, Closure $schemaClosure, array $options): ?ISchemaWrapper {
+		/** @var ISchemaWrapper $schema */
 		$schema = $schemaClosure();
-		$schema->dropTable('mail_classifiers');
+
+		if ($schema->hasTable('mail_classifiers')) {
+			$schema->dropTable('mail_classifiers');
+		}
+
 		return $schema;
 	}
 }
